backend/requirements.txt
adrf==0.1.12
amqp==5.3.1
anyio==4.13.0
asgiref==3.11.1
ast_serialize==0.4.0
async-property==0.2.2
attrs==26.1.0
billiard==4.2.4
black==26.3.1
celery==5.6.3
certifi==2026.4.22
channels==4.3.2
channels_redis==4.3.0
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.14
django-filter==25.2
django-parler==2.3
django-ratelimit==4.1.0
django-redis==6.0.0
django-stubs==6.0.4
django-stubs-ext==6.0.4
djangorestframework==3.16.1
djangorestframework-stubs==3.17.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
flake8==7.3.0
flower==2.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
idna==3.15
inflection==0.5.1
isort==8.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
librt==0.11.0
mccabe==0.7.0
msgpack==1.1.2
mypy==2.1.0
mypy_extensions==1.1.0
packaging==26.2
pathspec==1.1.1
pillow==12.1.1
platformdirs==4.9.6
prometheus_client==0.25.0
prompt_toolkit==3.0.52
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.4.1
pytz==2026.2
PyYAML==6.0.3
redis==7.1.1
referencing==0.37.0
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.5
tornado==6.5.5
types-PyYAML==6.0.12.20260510
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.7.0
